python/requirements.txt
websockets
Pillow